requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
openpyxl==3.1.2
WTForms==3.0.1
Flask-WTF==1.1.1
python-dateutil==2.8.2
cryptography==41.0.7